+++ src/components/Button/Button.tsx
@@ -27,6 +27,10 @@ interface IButtonProps {
     passiveColor?: 'green' | 'purple' | 'transparent' | 'transparent-green' | 'white';
     /** Hover color */
     hoverColor?: 'green' | 'purple' | 'transparent';
+    /** Border color */
+    border?: 'green' | 'transparent';
+    /** Font color */
+    fontColor?: 'white' | 'green';
     /** Click/press color */
     downColor?: 'dark' | 'transparent';
     /** Disabled color */
@@ -74,6 +78,14 @@ class Button extends React.Component<IButtonProps, {}> {
             'green',
             'purple',
             'transparent',
+        ]),
+        border: PropTypes.oneOf([
+            'green',
+            'transparent',
+        ]),
+        fontColor: PropTypes.oneOf([
+            'green',
+            'white',
         ]),
         downColor: PropTypes.oneOf([
             'dark',
@@ -129,8 +141,8 @@ class Button extends React.Component<IButtonProps, {}> {
     render() {
         const {
             sizeAll, sizeWide, sizeDesktop, sizeTablet, sizeMobile,
-            customView, passiveColor, hoverColor, downColor, disabledColor,
-            padding, width, margin, showSpinner, className, href, type,
+            customView, passiveColor, hoverColor, downColor, border, fontColor,
+            disabledColor, padding, width, margin, showSpinner, className, href, type,
             onClick, disabled, target, children,
         } = this.props;
         const ElementType = this.props.href ? 'a' : 'button';
@@ -148,6 +160,8 @@ class Button extends React.Component<IButtonProps, {}> {
                     'hover-color': !customView && hoverColor,
                     'down-color': !customView && downColor,
                     'disabled-color': !customView && disabledColor,
+                    'font-color': !customView && fontColor,
+                    'border': !customView && border,
                     padding,
                     width,
                     margin,